refactor(#113494): Extracting entries for lifetime processing to separate namespace.
