+++ src/libraries/Microsoft.Extensions.Http/src/LifetimeProcessing/Handlers/LifetimeTrackingHttpMessageHandler.cs
@@ -3,7 +3,7 @@
 
 using System.Net.Http;
 
-namespace Microsoft.Extensions.Http
+namespace Microsoft.Extensions.Http.LifetimeProcessing.Handlers
 {
     // This is a marker used to check if the underlying handler should be disposed. HttpClients
     // share a reference to an instance of this class, and when it goes out of scope the inner handler
